move new code to its own source file ApplicationStackHelper.cpp/.h
